Refactor pistesyotto: separate used for hakukohde and henkilo and keep common
