+++ src/components/koe-inputs-stateless.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, test, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
-import { KoeInputsStateless } from './koe-inputs';
+import { KoeInputsStateless } from './koe-inputs-stateless';
 import { ValintakoeInputTyyppi } from '@/lib/valintaperusteet/valintaperusteet-types';
 import {
   ValintakoeOsallistuminenTulos,
